rearrange file location, fix bug on client config always recreated

+++ shells/generateClient.sh
@@ -10,26 +10,32 @@ cd /opt
 if [[ ${OVPN_CLIENT_MODE} == "onlycert" ]]
 then
     echo "using certificate only"
-    bash certClientConfig.sh ${CLIENT_NAME}
+    bash ./helper/certClientConfig.sh ${CLIENT_NAME}
 fi
 
 if [[ ${OVPN_CLIENT_MODE} == "userpass" ]]
 then
     echo "using auth pam"
-    bash userpassClientConfig.sh ${CLIENT_NAME} ${CLIENT_PASS}
+    bash ./helper/userpassClientConfig.sh ${CLIENT_NAME} ${CLIENT_PASS}
 fi
 
 if [[ ${OVPN_CLIENT_MODE} == "userpasswithcert" ]]
 then
     echo "using auth pam"
-    bash userpassCertClientConfig.sh ${CLIENT_NAME} ${CLIENT_PASS}
+    bash ./helper/userpassCertClientConfig.sh ${CLIENT_NAME} ${CLIENT_PASS}
 fi
 
 
 if [[ "${OVPN_CLIENT_UNIQUE}" == 'true' ]]
 then
-    echo "" > ${OVPN_CLIENT_CCD}/${CLIENT_NAME}
-    echo $(date +"%F %T")" > config \"`ls ${OVPN_CLIENT_CCD}/${CLIENT_NAME}`\" created"
+    if [ -f "${OVPN_CLIENT_CCD}/${CLIENT_NAME}" ]
+    then
+        echo $(date +"%F %T")" > config \"`ls ${OVPN_CLIENT_CCD}/${CLIENT_NAME}`\" already exist"
+    else
+        echo "" > ${OVPN_CLIENT_CCD}/${CLIENT_NAME}
+        echo $(date +"%F %T")" > config \"`ls ${OVPN_CLIENT_CCD}/${CLIENT_NAME}`\" created"
+    fi
+    
 fi
 
 cd ${LAST_PWD}
